modified the font and interaction on the UI
