+++ package.json
@@ -19,7 +19,7 @@
     "embla-carousel-autoplay": "^8.5.2",
     "embla-carousel-fade": "^8.5.2",
     "embla-carousel-react": "^8.1.7",
-    "framer-motion": "^11.3.29",
+    "framer-motion": "^11.18.2",
     "lucide-react": "^0.417.0",
     "motion": "^12.0.1",
     "next": "15.1.6",
