+++ src/app/about/component/who-we-are.tsx
@@ -1,0 +1,19 @@
+import React from "react";
+
+export default function WhoWeAre() {
+  return (
+    <section className="bg-[#FFF3D6] py-20 max-w-screen-xl mx-auto px-8 lg:px-0">
+      <div className="text-center mb-12">
+        <h3 className="text-5xl text-center font-bold mb-12 text-primary">
+          Who We Are
+        </h3>
+        <p className="max-w-screen-lg mx-auto font-light text-lg leading-relaxed">
+          Our mission is to accelerate the transformation of healthcare in
+          Africa by building a future-ready workforce, fostering partnership for
+          innovation, and working with communities to co-design solutions that
+          enhance healthcare access, quality, and outcomes for everyone
+        </p>
+      </div>
+    </section>
+  );
+}
