+++ tailwind.config.ts
@@ -92,10 +92,13 @@ const config = {
       },
     },
     clipPath: {
-      mypolygon: "polygon(50% 0%, 100% 25%, 100% 75%, 50% 100%, 0% 75%, 0% 25%)",
+      mypolygon:
+        "polygon(50% 0%, 100% 25%, 100% 75%, 50% 100%, 0% 75%, 0% 25%)",
+      WebkitMaskImage:
+        "radial-gradient(circle, rgba(0,0,0,1) 60%, rgba(0,0,0,0) 100%)",
     },
   },
-  plugins: [require("tailwindcss-animate"), require('tailwind-clip-path'),],
+  plugins: [require("tailwindcss-animate"), require("tailwind-clip-path")],
 } satisfies Config;
 
 export default config;
